moved toggles unit test to gremlinapi
